Move tmpfile to conftest; update gcs_observer tests
